% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlearea.R
\name{circlearea}
\alias{circlearea}
\title{Calculate the area of a circle, given its radius}
\usage{
circlearea(r)
}
\arguments{
\item{r}{Radius of a circle}
}
\value{
Area of circle
}
\description{
Calculate the area of a circle, given its radius
}
\examples{
circlearea(1:5)

}
